buddypress-functions.php splitting: first pass
Now that `bp_locate_template()` is not preventing anymore loading templates into the administration we do not need anymore the bp-custom.php file. So to avoid confusions with the bp-custom.php file end users can add at the root of their `wp-content/plugins` directory we are removing this file.
All the classes and functions that it contained are moved to an `includes` subdirectory of the template pack and temporarly into a `functions.php` file.
The `BP_Nouveau` class has a new method to include required files: the `includes` method.

See #18