buddypress-functions.php splitting: final pass
Create a new file into the includes subdir for the common templates. It is called template-tags.php
Get rid of 2 bp-legacy deprecated functions: we can do it as we are building a new template pack so nobody is supposed to use bp-legacy functions with it.
Move the template tags into the newly created file.
Move the cover image function which is still to implement into the includes/functions.php file.

Fixes #18