Action buttons in loops: first step, the group of activity buttons
In BP Legacy to add an action button to each entry of the activity, we need to hook to "bp_activity_entry_meta" and your button is always after the delete link, which is not cool. I took the opportunity of our desicion to remove all do_action() hooks in template to improve this.

So the strategy i have chosen which i think can be apply to any buttons used in component loops is to create a new class "BP_Buttons_Group" to
build a group of "BP_Button". I am adding a "position" argument so that it is possible to choose where the custom button can be added.

A new template tag has been created "bp_nouveau_activity_entry_buttons()" and has replaced the hardcoded links into the "bp-nouveau/buddypress/actvity/entry.php" template.

The sort is only done once, as it is similar for each entries of the loop, and we simply update the previous buttons with parameters from the current activity.

The "bp_activity_entry_meta" is wrapped into the new template tag and a warning is displayed if WP_DEBUG mode is on to inform about the new way of adding custom action buttons for activity entries.

See #6