Activity group of action buttons, always create the same number of buttons.
As we are updating buttons to avoid sorting at each entry of the loop. We need to make sure the number of buttons created is always the same and remove the extra ones just before outputing the group of buttons. For instance an Activity post type cannot be commented and has 2 buttons, the following also had 2 buttons even if it could be commented. This commit is fixing this bug.

See #6