Introduce a new Activity Widget for the default front (or entire site)
Into the the dynamic_sidebar of the default front page, it will be possible to filter the activity query args using the bp_nouveau_activity_widget_query.
The output is using a new template /buddypress/activity/widget.php to easily allow overrides.
The widget allow the admin to select the type of activities to output.

See #19