Introduce 2 new BuddyPress dynamic sidebars for default single items front page
Members or Groups have their specific dynamic sidebars to add (BuddyPress) widgets to it.
To make sure the widget display is consistent with the displayed item, introduce filters that will be only used when inside the corresponding sidebars.
When no widgets are set a notice information helps the super admin to reach the right customizer section to choose the widgets.

See #19.

FYI: the primary nav widget code has been moved to the 'includes/classes.php' file.