Add new features to  Groups and Members default front pages
Thanks to this changeset https://buddypress.trac.wordpress.org/changeset/10905 we will now include a new section to the Groups default front page to display the widgets provided by Group extensions.
It is now possible to move the group description in this default front page instead of displaying it on every group page (being in the group header)
This 2 new features are activable/deactivable from the customizer.
For members default template, a new section has been added to let the user set his WordPress Bio info (the edit part still needs to be improved as for now it is a link to the WP Admin profile)
A new option has been added to the corresponding customizer section.
In the customizer, we are now hiding these new options if the complete use of the defautl front page has been disabled.

See #19