Share some groups/single/admin/ templates for the create screen of the new group
Adapt some templates and the function "bp_nouveau_group_manage_screen()" to make this happen.
Deprecate the hook "groups_custom_group_fields_editable" as it was deprecated in Legacy template and remove another 15 hooks from the create template (See #6)

NB: we can easily do this because we are not supporting the retired forum component.

See #42