The create group of buttons need to be displayed for any screens
Otherwise, as soon as there is a custom plugin we are stuck and cannot go to the next step or finish the group creation process.

So we are moving these buttons a bit lower into the bp_nouveau_group_manage_screen() function to make sure this trouble does not occur.

See #42