Give some love to the WP Profile of the displayed user
When the xProfile fields component is not active the WP Profile field is used. In Legacy the template is not great at all! It is using old profile info like Jabber, Aim or Yim.
In BP Nouveau we will use the wp_get_user_contact_methods() instead making sure if site owners are still using these old profile info they will be used, but if not they will not.
All custom contact methods will potentially be output. If some are to hide, a specific filter is available.
On a side note, thanks to this we are moving 5 more do_action hooks out of templates (see #6).