Standardize user feedback messages with template notices
Notice templates and user feedbacks are now using the same template part.
The function "bp_nouveau_get_user_feedback()" is where all the feedback messages are available. There are filters to edit them or add custom messages.
Feedbacks can include a dismiss button or not and fire hooks before and after or not.
4 more hooks inside the registration page has been moved out of the template (see #6).

Fixes #30