Improve the registration template
Use a unique template tag "bp_nouveau_signup_form()" to generate the signup form elements for the user and the blog signups.
Move 16 do_action hooks out of the templates (See #6).
Add aria-required and required attributes for the Browsers supporting it (eg: Google Chrome) to avoid the form to be posted if the fields required are empty.
Make sure the right protocol is used in the subdomains installs to improve the hardcoded "http://" of BP Legacy.