Bring Ajax into the single activity and improve favorites.
The single activity is the same as the directory activity except it has only one activity. The template part has been edited with this in mind.

Make sure deleting an activity will redirect the user to his home page.

Improve favorites and only remove the activity from the diplay when on the favorites scope (Directory or user).

See #26