Add Customizer settings to let end users choose directory loop layouts.
Members, Groups and Blogs can now be displayed in 1, 2 or 3 columns. A new section called "Directories layout" is available in the BP Template Pack panel of the customizer to let the admin choose their preferences about this display.

When the layout is 2 or 3 we will use the full avatar type.

See #52