Group invites UI: be more conservative.
So far the new Group Invites UI was allowing any member to invite any other one to join any group. The "All Members" tab was displayed by default and if the friends component was active, the "My Friends" one was displayed after.
Now, when the friends component is active it is the first nav displayed. The site owner can choose to use a filter to disable the "All Members" tab in this case. If the friends component is not active, it is still possible to invite any member to join any group.
If the friends component is active and the site owner decided to leave the "All members" tab and if the settings component is active. Users can now set their preferences about Group invites. To do so they can display the Group Invites settings page of their profile and choose to restrict group invites to their friends only.

Fixes #44