BP Buttons: BP_Buttons_group & Wrapper functions enhancements
Commit updates the nouveau button functions & classes to add neww params to build more flexible button types such as true 'button' element, input and updates the wrapper to container to make clear from element wrappers and allows it to be of type UL. $button arg arrays are updated for generic aray keys and values.

This upgrade requires the core bp button class to be updated , a patch is uploaded to the ticket linked to in the git ticket here.

Currently these updates will not function to pass the new args until the core patch is committed or local dev environments are patched N.B the functions will still work though as tested so far.

A little style update supports ul / li elements  in use and keeps items to inline display.

See #79