Update Messages star action
In `_accesories/messages/index.php` add aria-pressed states to the action buttons managing the star action.

Commit reverts previous change to using a `button` element to using an anchor due to JS non functionality in backbone - the button element should be used however.

N.B The Ajax functionality of this action is also non functional but where it breaks is undetermined.

See #161