Update 'Show All Comments' link JS
Commit just deals with the stray `addClass('hidden')` that still exist. For safety & as used to somehow show last comment if condition of 5 not met ( not clear on this part ). For safety retaining the classes but changing to `.bp-hidden` which we don't actually use but is checked on by the script sections. Added a removeClass to remove the class when showing all.

Additionally added aria role expanded = 'false' for initial show all link, and changed title attr to new `data-bp-tooltip`.

See #181 #161 & #163