Add BP Tooltips styles to bp-nouveau: First phase.
The first phase of BP tooltips styles adds the basic simple
tooltip which renders information from data-bp-tooltip to
bp-nouveau.

The arrow tips which uses the :before pseudo-element will be
introduced in the Second phase after we resolve conflict with
font icons which are added using the :before `content` as
well.

Resolution can be in the form of replacing font icons
with SVG icons per Issue #115 or via an existing or new
`span` element to hook the font icons to, etc.

See #163.