Revert out the 'outline' ruleset added.
The themes will be responsible for what happens with this property and links/form elements.
Currently we are not effecting this property but adding it is disruptive to many elements styled up for distinct visual appearance, if we are worried about focussed elements not being visually obvious when tabbed to then we would need to manage in a much more granular manner and probably with use of properties like background, although many elements already do have distinct visual cues  when they receive focus or hover.

See #162